RESPONSIVE EVVAI FOINALMENTE MACBOOK
