+++ src/components/ResponsivePages/LaptopResponsive/components/Maincontainer.tsx
@@ -1,13 +1,13 @@
 import React from "react";
 import SidePanel from "../../../SidePanel/SidePanel";
-import AboutMePanel from "./EditorSection/AboutMePanel";
+import { NavBarForPanel } from "./EditorSection/NavBarForPanel";
 import SideServerSection from "./TerminalBottom/SideServerSection";
 
 const MainContainer: React.FC = () => {
     return (
       <div className="flex h-[calc(100vh-34px-21px)]">
       <div className="w-[calc(100%-250px)] border-r border-borderColor">
-        <AboutMePanel />
+        <NavBarForPanel />
         <SideServerSection />
       </div>
         <SidePanel />
